chore: splits feature tests to separate files
As amount of test cases (groupped in Describe) started to grow we can
split a single file to several which are grouped per functionality.

Follow-up for #906.
