+++ tests/integration/features/serverless_feature_test.go
@@ -36,7 +36,7 @@ var _ = Describe("Serverless feature", func() {
 		// TODO rework
 		c, err := client.New(envTest.Config, client.Options{})
 		Expect(err).ToNot(HaveOccurred())
-		objectCleaner = envtestutil.CreateCleaner(c, envTest.Config, timeout, interval)
+		objectCleaner = envtestutil.CreateCleaner(c, envTest.Config, fixtures.Timeout, fixtures.Interval)
 
 		dsci = fixtures.NewDSCInitialization("default")
 		kserveComponent = &kserve.Kserve{}
@@ -83,7 +83,7 @@ var _ = Describe("Serverless feature", func() {
 				Expect(err).ToNot(HaveOccurred())
 				Expect(c.Create(context.TODO(), knativeServingCrdObj)).To(Succeed())
 
-				crdOptions := envtest.CRDInstallOptions{PollInterval: interval, MaxTime: timeout}
+				crdOptions := envtest.CRDInstallOptions{PollInterval: fixtures.Interval, MaxTime: fixtures.Timeout}
 				err = envtest.WaitForCRDs(envTest.Config, []*apiextensionsv1.CustomResourceDefinition{knativeServingCrdObj}, crdOptions)
 				Expect(err).ToNot(HaveOccurred())
 			})
@@ -271,7 +271,7 @@ var _ = Describe("Serverless feature", func() {
 				}
 
 				return nil
-			}).WithTimeout(timeout).WithPolling(interval).Should(Succeed())
+			}).WithTimeout(fixtures.Timeout).WithPolling(fixtures.Interval).Should(Succeed())
 		})
 
 		It("should not create any TLS secret if certificate is user provided", func() {
@@ -306,7 +306,7 @@ var _ = Describe("Serverless feature", func() {
 				}
 
 				return nil
-			}).WithTimeout(timeout).WithPolling(interval).Should(Succeed())
+			}).WithTimeout(fixtures.Timeout).WithPolling(fixtures.Interval).Should(Succeed())
 		})
 
 	})
