+++ tests/integration/features/servicemesh_feature_test.go
@@ -36,7 +36,7 @@ var _ = Describe("Service Mesh setup", func() {
 	BeforeEach(func() {
 		c, err := client.New(envTest.Config, client.Options{})
 		Expect(err).ToNot(HaveOccurred())
-		objectCleaner = envtestutil.CreateCleaner(c, envTest.Config, timeout, interval)
+		objectCleaner = envtestutil.CreateCleaner(c, envTest.Config, fixtures.Timeout, fixtures.Interval)
 
 		namespace := envtestutil.AppendRandomNameTo("service-mesh-settings")
 
@@ -175,7 +175,7 @@ var _ = Describe("Service Mesh setup", func() {
 
 				BeforeEach(func() {
 					smcpCrdObj = installServiceMeshCRD()
-					objectCleaner = envtestutil.CreateCleaner(envTestClient, envTest.Config, timeout, interval)
+					objectCleaner = envtestutil.CreateCleaner(envTestClient, envTest.Config, fixtures.Timeout, fixtures.Interval)
 					dsci = fixtures.NewDSCInitialization(namespace)
 
 					serviceMeshSpec = &dsci.Spec.ServiceMesh
@@ -243,7 +243,7 @@ var _ = Describe("Service Mesh setup", func() {
 							Expect(found).To(BeTrue())
 							return extensionProviders
 
-						}).WithTimeout(timeout).WithPolling(interval).Should(BeEmpty())
+						}).WithTimeout(fixtures.Timeout).WithPolling(fixtures.Interval).Should(BeEmpty())
 					})
 
 				})
@@ -260,7 +260,7 @@ func installServiceMeshCRD() *apiextensionsv1.CustomResourceDefinition {
 	Expect(yaml.Unmarshal([]byte(fixtures.ServiceMeshControlPlaneCRD), smcpCrdObj)).ToNot(HaveOccurred())
 	Expect(envTestClient.Create(context.TODO(), smcpCrdObj)).ToNot(HaveOccurred())
 
-	crdOptions := envtest.CRDInstallOptions{PollInterval: interval, MaxTime: timeout}
+	crdOptions := envtest.CRDInstallOptions{PollInterval: fixtures.Interval, MaxTime: fixtures.Timeout}
 	Expect(envtest.WaitForCRDs(envTest.Config, []*apiextensionsv1.CustomResourceDefinition{smcpCrdObj}, crdOptions)).To(Succeed())
 
 	return smcpCrdObj
